pyproject.toml
[tool.poetry]
name = "active-network-er"
version = "0.1.0"
description = ""
authors = ["Matteo <matteo.dora@ens.fr>"]

[tool.poetry.dependencies]
python = "3.10"
pandas = "^1.3.4"
numpy = "^1.21.4"
scipy = "^1.7.2"
networkx = "^2.6.3"
pyarrow = "^6.0.0"
cmocean = "^2.0"
tqdm = "^4.62.3"
matplotlib = "^3.4.3"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
Sphinx = "^4.3.0"
sphinx-rtd-theme = "^1.0.0"
pytest-cov = "^3.0.0"
ipykernel = "^6.5.0"
python-lsp-server = {extras = ["all"], version = "^1.2.4"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
